Added --no-seccomp argument to build-setup.sh
This works around a tricky problem we have building for 32bit arm
on a aarch64 machine. In order to build in the sandbox using seccomp
we would need to first cross compile a 32bit bubblewrap helper program
first.

For now however, it should be fine to build the 32bit bundles without
using seccomp and later run the binaries on a 32bit arm host using
a native 32bit flatpak and bubblewrap - running 32bit arm binaries
on an aarch64 target will again require a cross compiled bubblewrap.