Implemeting gpg-signing in build scripts
If build-setup.sh is given the new --gpg-sign argument, the passed
gpg key ID will be used to sign the build results.

It is currently left to the build machine admin to setup a GPG
key for signing in such a way that signing build results does not
prompt for a password.