Implemented git head tracking
Now the main buildSourceBuild() entry point takes a second "changed"
argument which is 1 of the git repo has changed or 0 if left unchanged.

Leave it up to the build source implementations whether they want to
rebuild or not.

Currently setup so that the autotools build sources, used to refresh
and install flatpak, ostree and libgsystem avoid rebuilds when unchanged,
and the base runtime which is not built with flatpak-builder also avoids
the rebuild (and notification) when sources are already up to date.