Enhanced IRC notification system
Now multiple IRC targets can be listed for the same module in the config,
also the targets must be configured with message type filters to select
which types of messages (regular/fail/success) are appropriate for the
given IRC target.

Default configuration now sends all messages to #flatpak-builds on
freenode and only build failures to #flatpak on the same network.