Enhanced app json builds to track changed state of app manifests
When fixing a manifest for an app, this will automatically trigger
a build on the build machines of the app repository, however all apps
will be rebuilt instead of only those for which the manifest has actually
changed.

This patch adds some tracking/cacheing for the continuous build mode
to only rebuild apps if their manifest has changed.