Added --headroom GB option to build scripts
This commit adds a feature which will check before launching
a build if a given amount of GB (default 10) are available before
launching the build.

When not available, first the exported ostree repo will be culled
to keep only the most recent commits of every existing ref, if this
does not liberate enough space for the build, then the logs are purged.
Finally the build will error out if not enough space is available.

This patch also cleans up the build scheduling in build-setup.sh
and ensures the LD_LIBRARY_PATH is setup to include the tooling
prefix when running the build from a cron job.